moved oracle to a primitive
