+++ crates/client/mempool/Cargo.toml
@@ -45,6 +45,7 @@ mp-receipt.workspace = true
 mp-state-update.workspace = true
 mp-transactions.workspace = true
 mp-utils.workspace = true
+mp-oracle.workspace = true
 
 # Starknet
 blockifier.workspace = true
