+++ crates/node/Cargo.toml
@@ -35,6 +35,7 @@ mp-block.workspace = true
 mp-chain-config.workspace = true
 mp-convert.workspace = true
 mp-utils.workspace = true
+mp-oracle.workspace = true
 
 # Starknet
 blockifier.workspace = true
