+++ crates/node/src/main.rs
@@ -12,11 +12,12 @@ use clap::Parser;
 use cli::{NetworkType, RunCmd};
 use mc_block_import::BlockImporter;
 use mc_db::DatabaseService;
-use mc_mempool::{oracle::Oracle, GasPriceProvider, L1DataProvider, Mempool};
+use mc_mempool::{GasPriceProvider, L1DataProvider, Mempool};
 use mc_metrics::MetricsService;
 use mc_rpc::providers::{AddTransactionProvider, ForwardToProvider, MempoolAddTxProvider};
 use mc_telemetry::{SysInfo, TelemetryService};
 use mp_convert::ToFelt;
+use mp_oracle::Oracle;
 use mp_utils::service::{Service, ServiceGroup};
 use service::{BlockProductionService, GatewayService, L1SyncService, RpcService, SyncService};
 use starknet_providers::SequencerGatewayProvider;
